requirements.txt
cython
loguru
lockfile
gdown
yacs
glfw
mujoco==2.3.7
dm-control==1.0.14
scikit-learn
gym
open3d
pillow
pyquaternion
pyyaml
scipy
tabulate
tqdm
transforms3d
trimesh
moviepy
roma
natsort
nlopt
chumpy
absl-py
lxml
easydict
multimethod
termcolor
shortuuid
fire
dm-env
hydra-core
hydra-ray-launcher
hydra-submitit-launcher
stable-baselines3==1.1.0
opencv-python
wandb
tensorboard
torch-scatter
grpcio-tools
